perf: use uint32 for TreeKind, token.Kind, and Position fields
Shrink TreeKind and token.Kind from uint (8B) to uint32 (4B). Shrink
Position.Line and Position.Column from int (8B) to uint32 (4B),
reducing Position from 16 to 8 bytes. This reduces the size of every
token.Token (two Position fields) from 72 to 56 bytes.

Measured formatting the entire samples-graphviz directory (998 DOT
files, from the graphviz repository) with:

  go test -bench=BenchmarkPrint -benchmem -benchtime=10s ./printer/

Before:
  Small   4.43 MB/s  21680 B/op  174 allocs/op
  Medium 14.41 MB/s  3825775 B/op  29293 allocs/op
  Large   8.46 MB/s  2515471462 B/op  17466394 allocs/op
After:
  Small   4.26 MB/s  20304 B/op  174 allocs/op
  Medium 14.68 MB/s  3599484 B/op  29293 allocs/op
  Large   8.54 MB/s  2377359558 B/op  17466394 allocs/op

Allocation count unchanged (17.5M) — same number of interface-boxed
objects but each is smaller. B/op dropped ~6% (2.5GB → 2.4GB for Large).

+++ scanner.go
@@ -127,7 +127,7 @@ func (sc *Scanner) next() {
 
 // pos returns the current position as a token.Position.
 func (sc *Scanner) pos() token.Position {
-	return token.Position{Line: sc.curLine, Column: sc.curColumn}
+	return token.Position{Line: uint32(sc.curLine), Column: uint32(sc.curColumn)}
 }
 
 func (sc *Scanner) skipWhitespace() {
